GitHub Pages improvements (#291)
* feat: Improve json creation

Use a script instead of commands in ghpages.yml to generate the all-events.json file.

The script has the following advantages:

- it makes it easier to generate all-events.json during development,
- it do not need to be updated each year: all years between 2007 and the current year are automatically considered,
- it simplifies the ghpages.yml workflow.

* feat: Build site on PR

Currently the site is build only on main. This commit change that to also triggers the build on PR (but without deploying it).

* feat: Make all-events.json available at the site root

This file is very interesting for those who want to access developers-conferences-agenda data in an API-like style.