Fix end date for multi days events
When we have a multi-days event, we need to set the
end date to be the day after.

Otherwise, the end date is considered at the very first hour of the day
which makes the last day not part of the event.

See https://github.com/scraly/developers-conferences-agenda/pull/672#issuecomment-1775600800